### 2020-10-27 #2790 @whedon
[PRE REVIEW]: A Repository to Simulate the Spread of COVID-19 in any Population with a Known Age-Structure, Made in Tübingen.
**Submitting author:** @vahid-sb (<a href="http://orcid.org/0000-0001-6544-7960">Vahid S. Bokharaie</a>)
**Repository:** <a href="https://github.com/vahid-sb/MiTepid_sim" target ="_blank">https://github.com/vahid-sb/MiTepid_sim</a>
**Version:** v0.0.5
