### 2020-11-04 #2790 @majensen
@oeco28 - how's it going in the Palouse? Could I interest you in reviewing [this software](#2790) for the [Journal of Open Source Software](https://joss.theoj.org)? Please have a look at [the review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and let me know in this thread. Or, if you have a colleague you could recommend, that would be wonderful. --@majensen
