### 2020-11-05 #2790 @oeco28
Dude!!
  All good! Staying safe. How are you? Long time no see! Let me see. Do you
mind if I take a look at the review process. when do you need it? I have
three kids under 6 at home and everything takes a bit longer than usual!

big hugs
Omar

On Wed, Nov 4, 2020 at 12:17 PM Mark Jensen <notifications@github.com>
wrote:

> @oeco28
> <https://urldefense.com/v3/__https://github.com/oeco28__;!!JmPEgBY0HMszNaDT!6WdPAO-tHp8EW-zPly15ipUIQxdWuWAaA1vzkJ3R57QK1TIGBwSUcsFFyTnw1q7udaU$>
> - how's it going in the Palouse? Could I interest you in reviewing this
> software <#m_-132020001114090690_2790> for the Journal of Open Source
> Software
> <https://urldefense.com/v3/__https://joss.theoj.org__;!!JmPEgBY0HMszNaDT!6WdPAO-tHp8EW-zPly15ipUIQxdWuWAaA1vzkJ3R57QK1TIGBwSUcsFFyTnwbt4RBUA$>?
> Please have a look at the review process
> <https://urldefense.com/v3/__https://joss.readthedocs.io/en/latest/reviewer_guidelines.html__;!!JmPEgBY0HMszNaDT!6WdPAO-tHp8EW-zPly15ipUIQxdWuWAaA1vzkJ3R57QK1TIGBwSUcsFFyTnwNN0zD1A$>
> and let me know in this thread. Or, if you have a colleague you could
> recommend, that would be wonderful. --@majensen
> <https://urldefense.com/v3/__https://github.com/majensen__;!!JmPEgBY0HMszNaDT!6WdPAO-tHp8EW-zPly15ipUIQxdWuWAaA1vzkJ3R57QK1TIGBwSUcsFFyTnwrnCBMng$>
>
