### 2020-11-09 #2790 @majensen
@AngryMaciek -- would you have time and inclination to review [this work](#2790) for JOSS? Please let me know - appreciated!
