### 2020-11-09 #2790 @AngryMaciek
@majensen : sure, I'd be very glad to experience the process from the other side.
I have just added myself to the reviewers list for JOSS, it's official :)
