### 2020-11-29 #2790 @majensen
Hi @btabibian - By all means - please have a look at the [reviewer doc](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). JOSS is maybe unusual in that the review is a lot more like a code review rather than a traditional paper review. Reviewers are encouraged to create issues in the author's repo, which the author will respond to with code improvements, for example.

If you're interested, that would be great. Since you and @vahid-sb may be colleagues, please have a look at the [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) and let us know here if there are any disclosures you may want to make.
