+++ README.rst
@@ -25,6 +25,8 @@ What this library has to offer can be summarized as follows:
 
 * The library comes with a set of visualizations tools that allows us to generate various plots necessary to study the effects of each policy in the evolution of the trajectories of different compartments. We can plot the trajectories per age group or as an aggregate for the whole population.
 
+* The contact rates used in the examples are the output of an optimisation scheme as implemented in `MiTepid_opt <https://github.com/vahid-sb/MiTepid_opt>`_ package. 
+
 Examples
 --------
 
