### 2020-12-04 #2790 @majensen
Sure @adi3 - can you have a look at my [comment](https://github.com/openjournals/joss-reviews/issues/2790#issuecomment-735321728) to @btabibian and let me know about any conflict of interest to describe? It won't necessary preclude a review.
Much appreciated!
