### 2020-12-04 #2790 @adi3
@majensen I confirm there is no conflict of interest. I am not associated with the author or the Max Planck Institute in any manner. By the way, I've reviewed for JOSS in the past so I am aware of the reviewing procedure and policies :)
