+++ paper.bib
@@ -1,26 +1,0 @@
- @article {Bok2020b,
-	author = {Bokharaie, Vahid S.},
-	title = {A Study on the Effects of Containment Policies and Vaccination on the Spread of SARS-CoV-2},
-	elocation-id = {2020.10.16.20213835},
-	year = {2020},
-	doi = {10.1101/2020.10.16.20213835},
-	publisher = {Cold Spring Harbor Laboratory Press},
-	URL = {https://www.medrxiv.org/content/early/2020/10/20/2020.10.16.20213835},
-	eprint = {https://www.medrxiv.org/content/early/2020/10/20/2020.10.16.20213835.full.pdf},
-	journal = {medRxiv}
-}
-
-
-@article {Bok2020a,
-	author = {Bokharaie, Vahid S.},
-	title = {A Stratified Model to Quantify the Effects of Containment Policies on the Spread of COVID-19},
-	elocation-id = {2020.04.10.20060681},
-	year = {2020},
-	doi = {10.1101/2020.04.10.20060681},
-	publisher = {Cold Spring Harbor Laboratory Press},
-	URL = {https://www.medrxiv.org/content/early/2020/04/16/2020.04.10.20060681},
-	eprint = {https://www.medrxiv.org/content/early/2020/04/16/2020.04.10.20060681.full.pdf},
-	journal = {medRxiv}
-}
-
-
