### 2020-12-04 #2880 @adi3
@vahid-sb thank you. As I go through the review checklist, I'll be requesting you to make several changes to your repository. This is to bring the repo and paper in line with JOSS standards. Most of them will be tiny, so please bear with me.

In the Installation section of your readme, you state "Or you can install it directly from the repository." Please provide the commands (starting from git clone...) in the readme for the user to the do this install from source.
