### 2020-12-05 #2880 @vahid-sb
Dear @adi3, I removed that sentence altogether since it has nothing to do with Mitepid_sim in particular and I do not want to turn the documentation into a resource for teaching the basics of python. And I am happy to make any changes that make the repo better and would be grateful for any such suggestion from you or anybody else, but I'd be also grateful if you'd be kind enough to tell me all such tiny changes in one batch rather than one by one.
