### 2020-12-05 #2880 @adi3
@vahid-sb Thank you for making the change. I'm afraid it is not possible for me to give you all the required changes in one batch. The process of reviewing is iterative and as I go through the checklist, some things are interrelated so we have to approach them in discrete steps. We are not aiming for an easy publication process here, but a high standard one.

Now, going through your paper, I see a major revision necessary. Your Summary section needs to be "a clear description of the high-level functionality and purpose of the software". Instead, your current Summary provides a background of the field and the problem. Of course, some background information is necessary to include, in order to set up the context. So perhaps you can remove the explanation of what SEIR/SIS models are. I see that a bunch of the content you need for the Summary section can be extracted from your current Statement of Need section, in particular, the places where you describe what the software does.

I hope this gives you some direction. Please let me know if I clarify anything further.

Lastly, what exactly is "Made in Tuebingen"? The way it is used in your sentences currently does not make sense to me.
