### 2020-12-08 vahid-sb/MiTepid_sim#2 @AngryMaciek
[JOSS Review] Copyright note in the LICENSE file
This issue is a part of the JOSS review (openjournals/joss-reviews#2880).

@vahid-sb
I am not sure if including your Copyright notice on top of the license text is the correct way to go about it.
Now it states:

>  Copyright (c) 2019 Vahid Samadi Bokharaie
>
>  GNU GENERAL PUBLIC LICENSE
>                        Version 3, 29 June 2007
>
>  Copyright (C) 2007 Free Software Foundation, Inc. <https://fsf.org/>
>  Everyone is permitted to copy and distribute verbatim copies
>  of this license document, but changing it is not allowed.

So who exactly holds the rights? Vahid Samadi Bokharaie or Free Software Foundation?
Also, please notice that it explicitly says that changing the license document is not allowed...

Could you please look deeper into the case, paste here some info on how to provide a copyright note for a GPL-3 work and adjust accordingly?
