### 2020-12-08 vahid-sb/MiTepid_sim#3 @AngryMaciek
[JOSS Review] Functionality + automated tests
This issue is a part of the JOSS review (openjournals/joss-reviews#2880).

Hi @vahid-sb ,
One of the tick boxes in our revision form asks if the functional claims of the software have been confirmed. Another one: "Are there automated tests or manual steps described so that the functionality of the software can be verified?"
I could not find any CI nor a walkthrough on how to carry out the simulations, however, I noticed a lot of scripts in the `examples` dir...

Could you please include this information? It would be really nice if you add a Continous Integration mechanisms with unit tests / simulations. Preferably, [GitHub Actions](https://github.com/features/actions), so that you keep everything related to the project in one place (however, I would not mind other CI solutions, if you prefer any).
