### 2020-12-08 vahid-sb/MiTepid_sim#4 @AngryMaciek
[JOSS Review] Community guidelines
This issue is a part of the JOSS review (openjournals/joss-reviews#2880).

Hi @vahid-sb,

From the reviewer's list:

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Please add these _explicite_. Also, it would be really good if you add a CODE_OF_CONDUCT for the project too :)
