### 2020-12-08 vahid-sb/MiTepid_sim#5 @AngryMaciek
[JOSS Review] Dependencies versions
This issue is a part of the JOSS review (openjournals/joss-reviews#2880).

Hi @vahid-sb,

I see that the installation is handled by `pip`, great!

However, in the README you do not provide versions for the packages your work rely on. You should, if you want to do reproducible research. You cannot guarantee that it will work with every possible combination of versions between all dependencies, right? I assume you did not test them all out. Also, you cannot forsee the future and can't tell if the next minor upgrade in `numpy` will not introduce a change in behaviour, thus breaking your code. However, what you can do is to specify the exact versions which you used for your development and testing. And for these you indeed _can_ guarantee the software will work.

So that is what I would ask you to do: please specify exact versions both in the README but **also in the setup.py** (so that users always get the same versions of dependencies during the installation).
