### 2020-12-08 vahid-sb/MiTepid_sim#6 @AngryMaciek
[JOSS Review] API missing
This issue is a part of the JOSS review (openjournals/joss-reviews#2880).

@vahid-sb
I cannot find any documentation on how to actually use `mitepid`... Did I skip it?
The core part of the repository (i.e. your package) should be well documented for the users to know how to use it in their reseach.
