### 2020-12-08 vahid-sb/MiTepid_sim#7 @AngryMaciek
[JOSS Review] Documentation: a statement of need
This issue is a part of the JOSS review (openjournals/joss-reviews#2880).

@vahid-sb

Following the review checklist for the repository's documentation:
> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Could you please include this information? I see that, in principle, one could extract it from the "Introduction" section in the README but a statement of need is more general and high-level. As far as I am concerned it would be OK just to add one more short section to that file.
