### 2020-12-08 vahid-sb/MiTepid_sim#8 @AngryMaciek
[JOSS Review] Documentation: Example usage
This issue is a part of the JOSS review (openjournals/joss-reviews#2880).

@vahid-sb

Following the reviewer's checklist:
>Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)?

At the moment when I am writting this post neither the functionality nor testing is documented for this project. Therefore I cannot tell what the examples from the directory `examples` actually do...
So, right now, my answer to the quoted qestion would be: _no_.
