### 2020-12-12 vahid-sb/MiTepid_sim#9 @AngryMaciek
[JOSS Review] Major paper revisions
This issue is a part of the JOSS review (openjournals/joss-reviews#2880).

Hi @vahid-sb,

I compiled and looked closer at your *paper.md*.
In my opinion it requires major revisions before we can proceed with the revisions checks.

My checklist points out:
> Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
>
> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
>
>State of the field: Do the authors describe how this software compares to other commonly-used packages?
>
> Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
>
> References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

The points I could raise now:
* there is no need to describe the types of models in such great detail. You can just provide a reference to a nice review on epidemiological models, preferably in a similar context :) Instead of that detailed description please include a motivation behind the software _"for a diverse, non-specialist audience"_.
* I notice the section "Statement of need" in the paper but it just renamed section "Introduction" from the README file, isn't it? In general, SoN and Introduction should contain distinct information. In the SoN you should put exactly: What is the problem the software solves, why is it important, who can use it, how does it approach the problem, what are the results, in what form, how is this useful etc.
* I see no information on the "state of the field". Granted, computational/statistical models are quite old and no reasonable person would ask you to recap 50 years of literature but I would like you to do some research on the current technology/tools/packages that people who carry out simulations in epidemiology use. I would like you to have a short section in the paper which presents your tool in the context of the current "standard" procedures and practices. This would be what I imagine as "state of the field".
* I would like you to improve the language in the text. The style fits a markdown file in the repository but it is a little too informal for a paper (for my taste). You should definietely rewrite phrases like: `we can stratify the into various relevant groups, for example, age-groups.`, `There are a few scripts`, ` what the code can do`, `you can look ` and the excessive use of word `quite`.
* I do not check references yet since this will for sure change. You need to include the references to epidemiological models as well as software packages (state of the field).
* In general, this publication looks more to me like a copy-paste of the _README.md_ file, which should not be the case. The publication should be more general and aimed _"for a diverse, non-specialist audience"_.

I hope these bullet points will guide you in the right direction.
Please let me know once you push version 2.0 of the publication.
