### 2020-12-18 #2880 @majensen
@adi3 thanks for the suggestion. Right now, I think we can leave the review open-  @vahid-sb, this is of course common for all of us especially at the year end. I appreciate your setting the expectations; a January timeframe is great.

(I sort of reserve pauses when an author has to drop the work for an indeterminate period. The problem with the "paused" flag is that it causes me to exclude the paper from my already tiny attention span!)
