### 2020-12-30 vahid-sb/MiTepid_sim#5 @AngryMaciek
After reading a little more on Python packages I have to come back here and correct my comment a little:

Freezing exact dependencies versions (like: `1.3.32`) in the `setup.py` for a package is not a good practice since it puts very strict constraints on all the users which would like to use a given piece of software. It is advised to use `>=` for versions which I believe makes sense.

On the other hand we **do** want to have a solid environment where the research software indeed works and our certainty asymptotically tends to 1 , right? Regardless of any future events there should be an always-working failsafe env.

So I would propose to manage versions with `>=` (and state that in the README too) but please also provide a very specific `requirements.txt` file with a recipe for an environment which you would usually use for you package (maybe generate it with `pip freeze`?)
