+++ examples/PLOS_One_Paper_Figures/Switching_Policies/Germany/Uncontained_then_switching_SD_uncon_start_30D/policy_details.txt
@@ -1,0 +1,12 @@
+
+        0.0    --->  Uncontained (R0 = 2.95)
+       30.0    --->  SD (R0 = 0.63)
+       60.0    --->  Uncontained (R0 = 2.95)
+       90.0    --->  SD (R0 = 0.63)
+      120.0    --->  Uncontained (R0 = 2.95)
+      150.0    --->  SD (R0 = 0.63)
+      180.0    --->  Uncontained (R0 = 2.95)
+      210.0    --->  SD (R0 = 0.63)
+      240.0    --->  Uncontained (R0 = 2.95)
+      270.0    --->  SD (R0 = 0.63)
+      300.0    --->  Uncontained (R0 = 2.95)
