+++ examples/PLOS_One_Paper_Figures/00_herd_immunity/Uncontained_SEIR_Germany_variable_x0R.txt
@@ -1,0 +1,24 @@
+
+          T_i   &   end_R   &  end_R-R_init   &   max_I 
+ $  \%$ &  $91.77\%$ &  $91.77\%$&  $14.35\%$ \\ \hline
+ $  \%$ &  $90.94\%$ &  $85.94\%$&  $12.78\%$ \\ \hline
+ $  \%$ &  $90.01\%$ &  $80.01\%$&  $11.24\%$ \\ \hline
+ $  \%$ &  $88.98\%$ &  $73.98\%$&  $9.76\%$ \\ \hline
+ $  \%$ &  $87.82\%$ &  $67.82\%$&  $8.33\%$ \\ \hline
+ $  \%$ &  $86.53\%$ &  $61.53\%$&  $6.95\%$ \\ \hline
+ $  \%$ &  $85.07\%$ &  $55.07\%$&  $5.65\%$ \\ \hline
+ $  \%$ &  $83.43\%$ &  $48.43\%$&  $4.44\%$ \\ \hline
+ $  \%$ &  $81.56\%$ &  $41.56\%$&  $3.31\%$ \\ \hline
+ $  \%$ &  $79.44\%$ &  $34.44\%$&  $2.31\%$ \\ \hline
+ $  \%$ &  $77.00\%$ &  $27.00\%$&  $1.44\%$ \\ \hline
+ $  \%$ &  $74.19\%$ &  $19.19\%$&  $0.74\%$ \\ \hline
+ $  \%$ &  $73.57\%$ &  $17.57\%$&  $0.62\%$ \\ \hline
+ $  \%$ &  $72.91\%$ &  $15.91\%$&  $0.51\%$ \\ \hline
+ $  \%$ &  $72.20\%$ &  $14.20\%$&  $0.41\%$ \\ \hline
+ $  \%$ &  $71.36\%$ &  $12.36\%$&  $0.32\%$ \\ \hline
+ $  \%$ &  $70.26\%$ &  $10.26\%$&  $0.24\%$ \\ \hline
+ $  \%$ &  $68.66\%$ &  $7.66\%$&  $0.17\%$ \\ \hline
+ $  \%$ &  $66.66\%$ &  $4.66\%$&  $0.11\%$ \\ \hline
+ $  \%$ &  $65.14\%$ &  $2.14\%$&  $0.06\%$ \\ \hline
+ $  \%$ &  $64.81\%$ &  $0.81\%$&  $0.02\%$ \\ \hline
+ $  \%$ &  $65.29\%$ &  $0.29\%$&  $0.01\%$ \\ \hline
