+++ examples/PLOS_One_Paper_Figures/Uncontained_then_single_policy/Germany/Uncontained_then_Elderly_self_isolate/policy_details.txt
@@ -1,0 +1,3 @@
+
+        0.0    --->  Uncontained (R0 = 2.95)
+       60.0    --->  Elderly_self_isolate (R0 = 2.37)
