+++ examples/PLOS_One_Paper_Figures/Compare_R0_Uncontained/SEIRGermany_variable_R0_eventual_I_R.txt
@@ -1,0 +1,22 @@
+
+          R0    &   end_R   &   max_I 
+        R0 2.05 &  $78.06\%$ &  $7.90\%$ \\ \hline
+        R0 2.15 &  $79.08\%$ &  $8.50\%$ \\ \hline
+        R0 2.25 &  $83.72\%$ &  $9.72\%$ \\ \hline
+        R0 2.35 &  $85.38\%$ &  $10.41\%$ \\ \hline
+        R0 2.45 &  $85.87\%$ &  $10.94\%$ \\ \hline
+        R0 2.55 &  $87.66\%$ &  $11.76\%$ \\ \hline
+        R0 2.65 &  $88.95\%$ &  $12.48\%$ \\ \hline
+        R0 2.75 &  $90.17\%$ &  $13.16\%$ \\ \hline
+        R0 2.85 &  $89.34\%$ &  $13.32\%$ \\ \hline
+        R0 2.95 &  $91.77\%$ &  $14.35\%$ \\ \hline
+        R0 3.05 &  $92.48\%$ &  $14.79\%$ \\ \hline
+        R0 3.15 &  $93.86\%$ &  $15.67\%$ \\ \hline
+        R0 3.25 &  $94.00\%$ &  $15.93\%$ \\ \hline
+        R0 3.35 &  $94.00\%$ &  $16.31\%$ \\ \hline
+        R0 3.45 &  $95.74\%$ &  $17.40\%$ \\ \hline
+        R0 3.55 &  $96.54\%$ &  $18.10\%$ \\ \hline
+        R0 3.65 &  $95.00\%$ &  $17.60\%$ \\ \hline
+        R0 3.75 &  $96.18\%$ &  $18.42\%$ \\ \hline
+        R0 3.85 &  $95.83\%$ &  $18.60\%$ \\ \hline
+        R0 3.95 &  $96.42\%$ &  $19.07\%$ \\ \hline
