+++ examples/PLOS_One_Paper_Figures/Uncontained_then_single_policy/Germany/00_policy_variable/SEIRGermany_variable_policy_eventual_I_R.txt
@@ -1,0 +1,10 @@
+
+         policy &   end_R   &   max_I 
+             UN &  $91.77\%$ &  $14.35\%$ \\ \hline
+             KI &  $81.85\%$ &  $11.02\%$ \\ \hline
+             EL &  $84.45\%$ &  $11.21\%$ \\ \hline
+           KIEL &  $72.73\%$ &  $7.98\%$ \\ \hline
+           KIOF &  $58.57\%$ &  $5.81\%$ \\ \hline
+           ADEL &  $34.18\%$ &  $4.29\%$ \\ \hline
+             SD &  $22.27\%$ &  $3.94\%$ \\ \hline
+             LD &  $16.41\%$ &  $3.84\%$ \\ \hline
