+++ examples/PLOS_One_Paper_Figures/Switching_Policies/Germany/Uncontained_then_switching_R0p9_R1p5_start_60D/policy_details.txt
@@ -1,0 +1,13 @@
+
+        0.0    --->  Uncontained (R0 = 2.95)
+       60.0    --->  1.5 (R0 = 1.50)
+       90.0    --->  0.9 (R0 = 0.90)
+      120.0    --->  1.5 (R0 = 1.50)
+      150.0    --->  0.9 (R0 = 0.90)
+      180.0    --->  1.5 (R0 = 1.50)
+      210.0    --->  0.9 (R0 = 0.90)
+      240.0    --->  1.5 (R0 = 1.50)
+      270.0    --->  0.9 (R0 = 0.90)
+      300.0    --->  1.5 (R0 = 1.50)
+      330.0    --->  0.9 (R0 = 0.90)
+      360.0    --->  Uncontained (R0 = 2.95)
