+++ examples/PLOS_One_Paper_Figures/Switching_Policies/Germany/Uncontained_then_switching_SD_uncon_start_60D/policy_details.txt
@@ -1,0 +1,12 @@
+
+        0.0    --->  Uncontained (R0 = 2.95)
+       60.0    --->  SD (R0 = 0.63)
+       90.0    --->  Uncontained (R0 = 2.95)
+      120.0    --->  SD (R0 = 0.63)
+      150.0    --->  Uncontained (R0 = 2.95)
+      180.0    --->  SD (R0 = 0.63)
+      210.0    --->  Uncontained (R0 = 2.95)
+      240.0    --->  SD (R0 = 0.63)
+      270.0    --->  Uncontained (R0 = 2.95)
+      300.0    --->  SD (R0 = 0.63)
+      330.0    --->  Uncontained (R0 = 2.95)
