### 2021-02-23 cf725f7
Script to generate PLOS One Paper Figures.
