+++ examples/script_vary_TI_in_uncontained_pop.py
@@ -1,110 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Fri Oct 23 22:32:19 2020
-
-@author: vbokharaie
-
-Studying what happens if T_Infectious varies.
-T_Infectious is the average time-period in individuals infected with SARS-CoV-2 are infectious.
-
-This script uses `mitepid` library to simulate the spread of SARS-CoV-2 in poulations of different countries.
-
-The optimised contact rates can be found in `miteipd/Optimised_B` folder.
-
-More details of the methodlogy can be found in:
-https://www.medrxiv.org/content/10.1101/2020.10.16.20213835v1
-
-"""
-
-
-if __name__ == '__main__':
-    # %%
-    import numpy as np
-    from pathlib import Path
-    import logging
-    logging.getLogger().setLevel(logging.CRITICAL)
-
-    import mitepid
-    from mitepid.policies import get_B_policy
-    from mitepid.utils import load_mat
-    from mitepid.epid_sim import epid_sim
-    from mitepid.plots import bplot_agg
-
-
-    # %% simulation general variables
-    dir_save_plots = Path(Path(__file__).resolve().parent, 'OUTPUTS', 'Compare_T_Infectious_Uncontained')
-
-    age_groups = ['[0-10]',  '[10-20]', '[20-30]', '[30-40]',
-                  '[40-50]', '[50-60]', '[60-70]', '[70-80]', '80+',]
-
-    module_path = Path(mitepid.__file__).parent
-    dir_B_opt = Path(module_path, 'Optimised_B')
-
-    t_end = 211
-    x00 = 1e-5
-    x0_vec=[x00, x00, x00, x00, x00, x00, x00, x00, x00,]
-
-    policy_name = 'Uncontained'
-    model_type = 'SEIR'
-    list_t_switch = [0,]
-    all_containment_list = ['Uncontained',]
-    country = 'Germany'
-    list_sol_I = []
-    list_sol_R = []
-    list_labels = ['Ti ' + str(x) + ' days' for x in np.arange(10) + 5]
-    for T_I in np.arange(10) + 5:
-        print('Infectious time = ', T_I)
-        model_type = 'SEIR'
-        file_Bopt = Path('B_opt_SEIR_Ti_' + str(T_I) + '_R0_2.95.mat')
-        file_B_opt = Path(dir_B_opt, file_Bopt)
-        Sigma = load_mat(file_B_opt, 'Sigma')
-        Gamma = load_mat(file_B_opt, 'Gamma')
-        B_opt_country = get_B_policy(file_B_opt, country, 'Uncontained', )
-        Ng = B_opt_country.shape[0]
-        assert len(age_groups)==9, 'Check the defined age-groups. Something is wrong there.'
-        dir_save = Path(dir_save_plots, 'T_I_variable', str(T_I))
-        epid_obj = epid_sim(model_type=model_type,
-                            B=B_opt_country,
-                            Gamma=Gamma,
-                            Sigma=Sigma,
-                            dir_save_plots_main=dir_save,
-                            country=country,
-                            policy_list=all_containment_list,
-                            policy_switch_times=list_t_switch,
-                            x0=x0_vec,
-                            t_end=t_end,
-                            str_policy=policy_name,
-                            group_labels=age_groups,)
-        sol_agg_I = epid_obj.sol_agg_dict['I']
-        sol_agg_R = epid_obj.sol_agg_dict['R']
[…]
