+++ examples/script_uncontained_then_single_policy.py
@@ -1,152 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Fri Oct 23 22:41:48 2020
-
-@author: vbokharaie
-
-Studying the effects of various Suppression Strategies on the spread of
-    SARS-CoV-2 in the population of Germany.
-
-This script uses `mitepid` library to simulate the spread of SARS-CoV-2 in poulations of different countries.
-
-The optimised contact rates can be found in `miteipd/Optimised_B` folder.
-
-More details of the methodlogy can be found in:
-https://www.medrxiv.org/content/10.1101/2020.10.16.20213835v1
-"""
-
-
-if __name__ == '__main__':
-    # %%
-    import numpy as np
-    from pathlib import Path
-    import logging
-    logging.getLogger().setLevel(logging.CRITICAL)
-
-    import mitepid
-    from mitepid.policies import get_B_policy
-    from mitepid.utils import load_mat
-    from mitepid.epid_sim import epid_sim
-    from mitepid.plots import bplot_agg
-
-
-    # %% simulation general variables
-    dir_save_plots = Path(Path(__file__).resolve().parent, 'OUTPUTS', 'Uncontained_then_single_policy')
-
-    age_groups = ['[0-10]',  '[10-20]', '[20-30]', '[30-40]',
-                  '[40-50]', '[50-60]', '[60-70]', '[70-80]', '80+',]
-
-    module_path = Path(mitepid.__file__).parent
-    dir_B_opt = Path(module_path, 'Optimised_B')
-
-    country = 'Germany'
-    t_end = 211
-    list_t_switch = [0]
-    x00 = 1e-5
-    x0_vec=[x00, x00, x00, x00, x00, x00, x00, x00, x00,]
-
-    #%%
-    model_type = 'SEIR'
-    file_Bopt = Path('B_opt_SEIR_Ti_5_R0_2.95.mat')
-    file_B_opt = Path(dir_B_opt, file_Bopt)
-    Sigma = load_mat(file_B_opt, 'Sigma')
-    Gamma = load_mat(file_B_opt, 'Gamma')
-    B_opt_country = get_B_policy(file_B_opt, country, 'Uncontained', )
-    Ng = B_opt_country.shape[0]
-    assert len(age_groups)==9, 'Check the defined age-groups. Something is wrong there.'
-
-    # %%
-    ref_obj = epid_sim(model_type=model_type,
-                        B=B_opt_country,
-                        Gamma=Gamma,
-                        Sigma=Sigma,
-                        dir_save_plots_main=dir_save_plots,
-                        country=country,
-                        policy_list=['Uncontained'],
-                        policy_switch_times=list_t_switch,
-                        x0=x0_vec,
-                        t_end=t_end,
-                        str_policy='Uncontained',
-                        group_labels=age_groups,)
-
-    list_t_switch = [0, 60]
-    policy_list = ['Uncontained',
-                   'Schools_closed',
-                   'Elderly_self_isolate',
-                   'Kids_Elderly_self_isolate',
-                   'Schools_Offices_closed',
-                   'Adults_Elderly_Self_isolate',
-                   'Social_Distancing',
-                   'Lockdown',
-                   ]
-    policy_labels = ['UN',
-                   'KI',
-                   'EL',
-                   'KIEL',
-                   'KIOF',
-                   'ADEL',
-                   'SD',
-                   'LD',
-                   ]
-    list_sol_I = []
-    list_sol_R = []
-    for policy in policy_list:
-        all_policies = ['Uncontained',]
-        all_policies.append(policy)
[…]

[…]
