### 2021-02-23 875f22b
Script to generate PLOS One Paper Figures.
