+++ mitepid/plots.py
@@ -20,7 +20,7 @@ def bplot_strat(t,
           if_plot_in_pc=True,
           cmap='Dark2',
           Ng=None,
-          policy_name_pos=0.75,
+          policy_name_pos=0.8,
           policy_legend=True,
           v_line=True,
           N_population=None):
@@ -87,16 +87,16 @@ def bplot_strat(t,
     # plt.style.use('ggplot')
     # mpl.rcParams['lines.linewidth'] = 3.0
     # mpl.rcParams['font.weight'] = 'bold'
-    font_size = 30
+    font_size = 40
     font = {'family' : 'DejaVu Sans',
                 'sans-serif' : 'Tahoma',
-                'weight' : 'regular',
+                'weight' : 'bold',
                 'size'   : font_size}
     mpl.rc('grid', color='#316931', linewidth=1, linestyle='dotted')
     mpl.rc('font', **font)
     mpl.rc('lines', lw=3,)
-    mpl.rc('xtick', labelsize=font_size)
-    mpl.rc('ytick', labelsize=font_size)
+    # mpl.rc('xtick', labelsize=font_size)
+    # mpl.rc('ytick', labelsize=font_size)
 
     if cmap == 'viridis_r':
         colors = pl.cm.viridis_r(np.linspace(0,1,Ng))
@@ -121,28 +121,28 @@ def bplot_strat(t,
         sol = sol*N_population
 
     fig, ax = plt.subplots(1, 1, figsize=(24,18))
-    ax.set_facecolor('0.95')
+    # ax.set_facecolor('0.95')
     fig.subplots_adjust(bottom=0.15, top=0.92, left=0.1, right = 0.85)
     for cc in np.arange(Ng):
         # my_label = 'x'+str(cc+1).zfill(2)
 
         if not labels==['']:
             ax.plot(t, sol[:, cc] * y_ax_scale, label=labels[cc], color = colors[cc], alpha=0.98)
-            ax.legend(bbox_to_anchor=(1.2, 1.0), prop={'size': 24})
+            ax.legend(bbox_to_anchor=(1.2, 1.0), prop={'size': 24, 'weight': 'bold'})
         else:
             ax.plot(t, sol[:, cc] * y_ax_scale, color = colors[cc], alpha=0.98)
 
         if ylim:
             ax.set_ylim(ylim)
-        ax.set_xlabel('\nTime (days)')
+        ax.set_xlabel('\nTime (days)', fontweight='bold')
         ax.set_xticks(np.arange(0, t[-1], step=30))
         plt.xticks(rotation=90)
-        ax.set_ylabel(ylabel)
+        ax.set_ylabel(ylabel, fontweight='bold')
         ylim_max = ax.get_ylim()[1]
     for idx1, xc in enumerate(list_vl):
         if v_line:
             ax.axvline(x=xc, color='r', linestyle='--', linewidth=1)
-        bbox = {'fc': '0.9', 'pad': 4, 'alpha': 0.3}
+        bbox = {'fc': '0.92', 'pad': 4, 'alpha': 0.3}
         props = {'ha': 'center', 'va': 'center', 'bbox': bbox,}
         if policy_legend and not list_all_policies is None:
             my_text = list_all_policies[idx1]
@@ -156,17 +156,18 @@ def bplot_strat(t,
             else:
                 policy_label_loc = 0.3*ylim_max
             xc_plot = xc+10*t[-1]/500
-            ax.text(xc_plot, policy_label_loc, my_text, props, rotation=90, color='k', alpha=0.7)
-    fig.suptitle(suptitle, fontsize=16, fontweight='bold')
+            ax.text(xc_plot, policy_label_loc, my_text, props,
[…]
