### 2021-06-07 #2880 @vahid-sb
Hi @majensen, thanks for your message and I hope life is treating you well. I used this library to do a study whose results I have already published. The paper itself took a lot of mental energy, in particular since I had to do it all alone. And although afterwards, I tried to force myself into editing this library, I couldn't find the required mental energy. And given what is going on in my professional life right now, I find it very unlikely that I can do the required editing for this library any time soon. Hence, I think it is better to withdraw this submission. I thank you and the reviewers for your time and comments. And please let me know what should I do to formally withdraw the review request.
